Tell parser to parse comments for issue #2
This commit reworks how the header is rewritten for adding
and removing imports. Previously the header was calculated
and then modified. The new behavior is to pass full source
to parser.ParseFile, with the ImportsOnly and ParseComments
flags set. This removes the need for the Header method and
logic, as well as parsing any comments.

The original size is calculated by the .End() of the last
token, or the .End() of the last comment, whichever is greater.
This can probably be simplified further.

- modimport methods now use parser flags properly
- Header method removed, handled by ImportsOnly flag
- AddImport and RemoveImport now return old-size
- ParseComments flag added to directly address issue